added some comments and refactored the file structure
