+++ app/client/Game of Life/LifeView.js
@@ -1,3 +1,6 @@
+/**
+This component renders our game of life to the DOM using a canvas element and allows users to toggle cells by clicking on them
+**/
 
 import React from 'react';
 
@@ -14,32 +17,29 @@ const LifeView = React.createClass({
 	},
 
 	componentDidMount() {
-		console.log('lifeview has mounted');
 		this.setState({container: document.getElementById('canvas').getContext('2d')});
 	},
 
 	componentDidUpdate() {
-		console.log('lifeview has updated');
 		this.draw();
 	},
 
-	click(e) {
-		this.props.toggleCell(Math.floor(e.nativeEvent.offsetY / this.state.height), Math.floor(e.nativeEvent.offsetX / this.state.width));
+	clickCell(e) {
+		this.props.toggleCell(Math.floor(e.nativeEvent.offsetY / this.state.height), 	//row and column respectively
+													Math.floor(e.nativeEvent.offsetX / this.state.width));
 	},
 
 	draw() {
 		let ctx = this.state.container;
-		// let ctx = document.getElementById('canvas').getContext('2d');
 
-		ctx.clearRect(0, 0, this.state.width * this.state.columns, this.state.height * this.state.rows);
-		console.log('drawing, first row is', this.state.cells[0]);
+		ctx.clearRect(0, 0, this.state.width * this.state.columns, this.state.height * this.state.rows);			//clear board
 
 		for (let j = 0; j < this.state.rows; j++) {
 			for (let i = 0; i < this.state.columns; i++) {
 				ctx.fillStyle = 'black';
 				ctx.fillRect(i * this.state.width, j * this.state.height, this.state.width, this.state.height);
 
-				ctx.fillStyle = this.state.cells[j][i] ? 'red' : 'white';
+				ctx.fillStyle = this.state.cells[j][i] ? 'red' : 'white';																													//red if alive, white if dead
 				ctx.fillRect(i * this.state.width + 1, j * this.state.height + 1, this.state.width - 2, this.state.height - 2);
 			}
 		}
@@ -49,7 +49,7 @@ const LifeView = React.createClass({
 		return (
 			<canvas id='canvas' height={this.state.height * this.state.rows}
 			width={this.state.width * this.state.columns}
-			onClick={this.click}></canvas>
+			onClick={this.clickCell}></canvas>
 		);
 	}
 });
